+++ app/src/main/java/ru/harrier55/project/filmography/domain/repo/CardFilmRepoImpl.kt
@@ -1,6 +1,9 @@
 package ru.harrier55.project.filmography.domain.repo
 
 import ru.harrier55.project.filmography.data.*
+import ru.harrier55.project.filmography.data.webconnection.OnRequestCompleteListener
+import ru.harrier55.project.filmography.data.webconnection.WebConnectionOkHttp
+import ru.harrier55.project.filmography.data.webconnection.WebConnectionRetrofit
 import ru.harrier55.project.filmography.domain.entities.CardFilmEntity
 
 typealias CacheListFilmsListener = (listFilms: List<CardFilmEntity>) -> Unit
@@ -18,7 +21,6 @@ class CardFilmRepoImpl(): IFilmRepo {
     }
 
     override fun updateCardFilm() {
-        TODO("Not yet implemented")
         notifyChanges()
     }
