+++ app/src/main/java/ru/harrier55/project/filmography/domain/repo/IFilmRepo.kt
@@ -1,7 +1,7 @@
 package ru.harrier55.project.filmography.domain.repo
 
 
-import ru.harrier55.project.filmography.data.OnRequestCompleteListener
+import ru.harrier55.project.filmography.data.webconnection.OnRequestCompleteListener
 import ru.harrier55.project.filmography.domain.entities.CardFilmEntity
 
 interface IFilmRepo {
